refactor(dis_responder): extract HTML presentation helpers into dis_html.py
dis_responder.py was ~13k lines. Move the pure, presentation-only helpers
(no DB/request/app coupling) to a new dis_html.py and re-import them so every
call site stays unchanged.

Moved: render_table/cell/fcell/safe (+ Safe/_Cell str subclasses), stat_cards,
tiny_badge, render_warning, oa_status_rank (+ OA_STATUS_ORDER), doi_link,
make_link, create_downloadable/dloop (+ DOWNLOAD_ICON), year_pulldown,
generate_navbar (+ NAV). The 14 names called by routes are re-exported via
`from dis_html import (...)`; helpers used only internally (_render_cell,
random_string, generate_navbar_items) and the moved constants stay private to
dis_html.

No behavior change: add_jrc_fields and anything touching DB/request/routes stay
put (add_jrc_fields calls the moved doi_link via the re-export). dis_responder
drops 13043 -> 12601 lines. Verified: both parse, all moved units execute
standalone, no dangling refs, and live pages (tables, navbar, stat cards, OA
detail, DOI detail, POST drill) all return 200. Patch bump to 119.24.1.
